Move for-client formatDate() to app/components/util.ts
